pyproject.toml
[tool.poetry]
name = "aiops-engine"
version = "0.1.0"
description = ""
authors = ["Akhilesh Jain <jainakhilesh080@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
lancedb = "^0.16.0"
azure-mgmt-monitor = "^6.0.2"
azure-identity = "^1.19.0"
azure-mgmt-automation = "^1.0.0"
llama-index = "^0.12.1"
boto3 = "^1.35.66"
easydict = "^1.13"
adlfs = "^2024.7.0"
tenacity = "^8.2.0, <8.4.0 || >8.4.0, <9.0.0"
openai = "^1.55.1"
instructor = "1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
